Series of bugfixes. Files uploaed without an extension are now examined to determine file type. If file type cannot be determined, error is presented to user about file extension. Error in script "FASTQ_1_trimmed.sh" for fastq files with no errors which resulted in deletion of data file has been corrected.

+++ dev/sh/FASTQ_1_trimming.sh
@@ -22,7 +22,6 @@ else
 		echo "# ";
 		echo "# Trimming of unbalanced read pairs has already been run on this datsaet.";
 		echo "# ";
-		echo;
 		exit 0;   # exit with no error.
 	else
 		## FASTQ format per line, repeating.
@@ -32,55 +31,55 @@ else
 		# echo "quality"
 
 		# number of lines of each file.
-		echo -e "\tTotal number of lines in file:";
+		echo "#\tTotal number of lines in file:";
 		length_full1=$(wc -l $1 | awk '{print $1}');
-		echo -e "\t\tFile1 : "$length_full1;
-		echo;
+		echo "#\t\tFile1 : "$length_full1;
 
 		# modulus of the number of lines by 4, as the fastq format is in blocks of 4 lines.
 		# this will give us the number of extra lines in each file.
-		echo -e "\tLines at end of each file suggestive of a cropped entry:";
+		echo "#\tLines at end of each file suggestive of a cropped entry:";
 		length_extra1=$(expr $length_full1 % 4);
-		echo -e "\t\tFile1 : "$length_extra1;
-		echo;
+		echo "#\t\tFile1 : "$length_extra1;
 
-		# the number of properly formated lines per file.
-		echo -e "\tNumber of properly formated lines per file:";
-		length_base1=$(expr $length_full1 - $length_extra1);
-		echo -e "\t\tFile1 : "$length_base1;
-		echo;
-
-		# find the residue lengths per file.
-		# this is the number of lines of incomplete and extra reads.
-		echo -e "\tNumber of lines at the end of file to be removed:";
-		length_residue1=$length_extra1;
-		echo -e "\t\tFile1 : "$length_residue1;
-		echo;
-
-		findStr=".fastq";
-		replaceStr=".trimmed.fastq";
-		trimmedName1=$(echo $1 | sed -e "s/$findStr/$replaceStr/g");
-
-		# make copy of the original read file which are trimmed to valid length.
-		echo -e "\tMaking trimmed valid fastq files:";
-		echo -e "\t\t"$1" -> "$trimmedName1;
-		cat $1 | head -$length_base1 > $trimmedName1;
-		echo;
-
-		# make residue file containing any incomplete or extra read lines.
-		echo -e "\tMaking file containing residual lines:";
-		if [ "$length_residue1" -gt 0 ]
+		# If the number of extra lines is zero, then nothing needs to be done.
+		# If the number of extra lines is more than zero, then create a trimmed file to replace the original.
+		if [ "$length_extra1" -gt 0 ]
 		then
-			echo -e "\t\t"$1" -> "$residueName1;
-			cat $1 | tail -$length_residue1 > $residueName1;
-		fi
-		echo;
+			# the number of properly formated lines per file.
+			echo "#\tNumber of properly formated lines per file:";
+			length_base1=$(expr $length_full1 - $length_extra1);
+			echo "#\t\tFile1 : "$length_base1;
 
-		echo -e "\tReplacing original files with trimmed versions.";
-		# cleanup original raw files.
-		rm $1;
-		# rename produced valid FASTQ files to original file names.
-		mv $trimmedName1 $1
-		echo;
+			# find the residue lengths per file.
+			# this is the number of lines of incomplete and extra reads.
+			echo "#\tNumber of lines at the end of file to be removed:";
+			length_residue1=$length_extra1;
+			echo "#\t\tFile1 : "$length_residue1;
+
+			findStr=".fastq";
+			replaceStr=".trimmed.fastq";
+			trimmedName1=$(echo $1 | sed -e "s/$findStr/$replaceStr/g");
+
+			# make copy of the original read file which are trimmed to valid length.
+			echo "#\tMaking trimmed valid fastq files:";
+			echo "#\t\t"$1" -> "$trimmedName1;
+			cat $1 | head -$length_base1 > $trimmedName1;
+
+			# make residue file containing any incomplete or extra read lines.
+			echo "#\tMaking file containing residual lines:";
+			if [ "$length_residue1" -gt 0 ]
+			then
+				echo "#\t\t"$1" -> "$residueName1;
+				cat $1 | tail -$length_residue1 > $residueName1;
+			fi
+
+			echo "#\tReplacing original files with trimmed versions.";
+			# cleanup original raw files.
+			rm $1;
+			# rename produced valid FASTQ files to original file names.
+			mv $trimmedName1 $1
+		else
+			echo "#\tFile trimming does not need performed on this dataset.";
+		fi
 	fi
 fi
